worker/requirements.txt
redis==5.2.1
psycopg[binary]==3.2.4
psycopg-pool==3.2.4
pydantic==2.10.6
pydantic-settings==2.7.1
Pillow==11.1.0
pyvips==2.2.3
imagehash==4.3.1
onnxruntime==1.20.1
huggingface-hub==0.30.1
pandas==2.2.3
numpy==2.2.3
